requirements.txt
streamlit
openai
faiss-cpu
langchain
tqdm
python-dotenv
